+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/Normal.cpp
@@ -32,8 +32,8 @@ Normal::Normal()
 {
 }
 
-void	Normal::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	Normal::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
