+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/SpecularControl.cpp
@@ -32,8 +32,8 @@ SpecularControl::SpecularControl()
 {
 }
 
-void	SpecularControl::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	SpecularControl::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
