+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Cameras/CameraRange.cpp
@@ -32,8 +32,8 @@ CameraRange::CameraRange()
 {
 }
 
-void	CameraRange::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	CameraRange::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 8)
 	{
