+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Box.cpp
@@ -32,10 +32,10 @@ Box::Box()
 {
 }
 
-void	Box::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Box::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
