+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Torus.cpp
@@ -32,10 +32,10 @@ Torus::Torus()
 {
 }
 
-void	Torus::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Torus::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen == 0)
 	{
