+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Lights/LightData.cpp
@@ -32,8 +32,8 @@ LightData::LightData()
 {
 }
 
-void	LightData::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	LightData::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 20)
 	{
