+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Lights/SpotLight.cpp
@@ -44,8 +44,8 @@ SpotLight::SpotLight()
 {
 }
 
-void	SpotLight::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	SpotLight::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 52)
 	{
