+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Miscellaneous/ProcessUnknownType.cpp
@@ -39,15 +39,15 @@
 	@param		inTypeName		Class name of the type if known, else NULL.
 */
 void	ProcessUnknownType( uint32_t inType,
-							uint32_t inStartOffset,
-							uint32_t inEndOffset,
+							size_t inStartOffset,
+							size_t inEndOffset,
 							const char* inTypeName )
 {
 	Controller* cntl = Controller::Get();
 	
 	cntl->OutStream() << cntl->Indent() << "UnknownBinary (\n";
 	
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	cntl->OutStream() << cntl->Indent(1) << (int32_t)inType << " " <<
 		dataLen << " " <<
