+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/PickPartsStyle.cpp
@@ -43,8 +43,8 @@ PickPartsStyle::PickPartsStyle()
 {
 }
 
-void	PickPartsStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	PickPartsStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
