+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/Scale.cpp
@@ -32,10 +32,10 @@ Scale::Scale()
 {
 }
 
-void	Scale::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Scale::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 
 	if (dataLen != 12)
 	{
