+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/ShaderUVTransform.cpp
@@ -32,9 +32,9 @@ ShaderUVTransform::ShaderUVTransform()
 {
 }
 
-void	ShaderUVTransform::Process( uint32_t inStartOffset, uint32_t inEndOffset )
+void	ShaderUVTransform::Process( size_t inStartOffset, size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen != 36)
 	{
